Avoid collisions when moving players
using a map function and passing the same board to each user meant
that 2 users may try to move into the same open spot during the same
iteration. To combat this, we get positions for each player
individually, then update the board with that players position before
moving on to the next player.