Fix players not moving correctly when speed greater than 1
When speed was greater than one, the sense was failing to move the
player correctly. The player would only be moved one spot from their
original position. This was because when creating a map of players who
moved within a turn. We would always set the player in the map. This
means we'd only ever get the first position of the player, on subsequent
moves of the player (if the players speed was greater than 1). This
meant no matter what parameters you added to the player, they would
always go down to 1 speed.